refactor: put endpoints in a `service` module
